pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gis-train"
version = "0.1.0"
description = "Train AI models for crop classification from Sentinel-2 imagery over Uzbekistan / Central Asia."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "ai-train-on-gis-data contributors" }]
keywords = ["geospatial", "gis", "remote-sensing", "sentinel-2", "crop-classification", "pytorch", "torchgeo"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "torchgeo>=0.6",
    "pytorch-lightning>=2.2",
    "torchmetrics>=1.3",
    "numpy>=1.26",
    "rasterio>=1.3",
    "geopandas>=0.14",
    "shapely>=2.0",
    "pyproj>=3.6",
    "pystac-client>=0.7",
    "planetary-computer>=1.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "scikit-learn>=1.4",
    "pandas>=2.2",
    "tqdm>=4.66",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "pre-commit>=3.6",
]
notebooks = [
    "jupyterlab>=4.0",
    "matplotlib>=3.8",
    "folium>=0.15",
    "ipywidgets>=8.1",
]
cropharvest = [
    # Optional: full CropHarvest Python API. Only pulled in on demand
    # because it has heavy transitive deps.
    "cropharvest>=0.7",
]

[project.urls]
Homepage = "https://github.com/bahromiddin-0256/ai-train-on-gis-data"
Issues = "https://github.com/bahromiddin-0256/ai-train-on-gis-data/issues"

[project.scripts]
gis-train = "gis_train.train:cli"
gis-evaluate = "gis_train.evaluate:cli"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["gis_train*"]

[tool.setuptools.package-data]
gis_train = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "N812",  # lowercase imported as non-lowercase (F as torch.nn.functional)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011", "N802"]
"notebooks/**" = ["E402", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning:torchgeo.*",
    "ignore::UserWarning:pytorch_lightning.*",
]

[tool.coverage.run]
branch = true
source = ["gis_train"]

[tool.coverage.report]
show_missing = true
skip_covered = true
